backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
pydantic==2.9.2
httpx==0.27.2
python-multipart==0.0.12

# Semantic matching (UCCF engine)
sentence-transformers
spacy

# Environment / secrets
python-dotenv>=1.0.0

# AutoAlign Policy Factory & Converter
openai>=1.30.0
python-docx>=1.1.0
aiofiles>=23.2.1
mammoth>=1.6.0
scikit-learn>=1.4.0
numpy>=1.26.0
lxml>=4.9.0

# DOCX → PDF conversion (requires Microsoft Word on Windows)
docx2pdf>=0.1.8

# Diagram rendering (swimlane + flowchart via mermaid.ink)
matplotlib>=3.8.0
requests>=2.31.0
